requirements.txt
Django>=3.0,<4.0
psycopg2>=2.9.6
psycopg2-binary>=2.9.6
djangorestframework==3.12.4